requirements.txt
numpy
pandas
dataclasses-json
sqlalchemy-iris
flask
requests
traceback2
mcp==1.8.0
uvicorn==0.34.2
iris-embedded-python-wrapper